+++ compiler/src/test/fixtures/basic/output/java-cdi-client/com/example/basic/EventServiceImpl.java
@@ -7,6 +7,7 @@ import javax.annotation.Resource;
 import javax.enterprise.context.ApplicationScoped;
 import javax.enterprise.inject.Typed;
 import javax.transaction.Transactional;
+import javax.ws.rs.ProcessingException;
 import javax.ws.rs.client.ClientBuilder;
 import javax.ws.rs.client.Entity;
 import javax.ws.rs.client.Invocation;
@@ -33,6 +34,8 @@ public class EventServiceImpl implements EventService {
       } else {
         throw new GraphQLException( "Error invoking GraphQL endpoint. HTTP Status: " + response.getStatusInfo() );
       }
+    } catch ( final ProcessingException pe  ) {
+      throw new GraphQLException( "Communication error invoking the GraphQL endpoint.", pe );
     }
   }